Add folium markdown support to rf_ipython module
